+++ ta4j-core/src/main/java/org/ta4j/core/bars/TickBarBuilderFactory.java
@@ -32,7 +32,7 @@ public class TickBarBuilderFactory implements BarBuilderFactory {
 
     private final boolean realtimeBars;
     private final int tickCount;
-    private TickBarBuilder barBuilder;
+    private transient TickBarBuilder barBuilder;
 
     /**
      * Constructor.